[refactor] #79 영수증 리뷰 악의적 접근 차단
